🩹 use uint256 until casted in vault

+++ src/Factory.sol
@@ -189,7 +189,7 @@ contract Factory is ERC721, Ownable {
         IVault(vault).initialize(
             msg.sender,
             vaultDetails[currentVaultVersion].registryAddress,
-            uint8(numeraire),
+            numeraire,
             numeraireToStable[numeraire],
             vaultDetails[currentVaultVersion].stakeContract,
             vaultDetails[currentVaultVersion].interestModule
